test-agent/requirements.txt
google-generativeai>=0.3.0
solders>=0.18.0
solana>=0.30.0
requests>=2.31.0
python-dotenv>=1.0.0
base58>=2.1.0

